(feature) 파일 이름 변경 및 임시 input 지점 설정
