+++ cs/CMakeLists.txt
@@ -223,7 +223,6 @@ set(PARSER_SOURCES
   ${PARSER_DIR}/view_transform.c
   ${PARSER_DIR}/xasl_regu_alloc.cpp
   ${PARSER_DIR}/xasl_generation.c
-  ${PARSER_DIR}/predication/parser_expr_compiler.cpp
   ${PARSER_DIR}/func_type.cpp
   )
 set(PARSER_HEADERS
@@ -231,7 +230,6 @@ set(PARSER_HEADERS
   ${PARSER_DIR}/parse_type.hpp
   ${PARSER_DIR}/parser_allocator.hpp
   ${PARSER_DIR}/xasl_regu_alloc.hpp
-  ${PARSER_DIR}/predication/parser_expr_compiler.hpp
   )
 
 set(OPTIMIZER_SOURCES
@@ -479,7 +477,6 @@ SET_SOURCE_FILES_PROPERTIES(
   ${PARSER_DIR}/type_checking.c
   ${PARSER_DIR}/view_transform.c
   ${PARSER_DIR}/xasl_generation.c
-  ${PARSER_DIR}/predication/parser_expr_compiler.cpp
   ${OPTIMIZER_SOURCES}
   ${QUERY_SOURCES}
   ${OBJECT_SOURCES}
