+++ sa/CMakeLists.txt
@@ -235,14 +235,12 @@ set(PARSER_SOURCES
   ${PARSER_DIR}/view_transform.c
   ${PARSER_DIR}/xasl_regu_alloc.cpp
   ${PARSER_DIR}/xasl_generation.c
-  ${PARSER_DIR}/predication/parser_expr_compiler.cpp
   )
 set(PARSER_HEADERS
   ${PARSER_DIR}/func_type.hpp
   ${PARSER_DIR}/parse_type.hpp
   ${PARSER_DIR}/parser_allocator.hpp
   ${PARSER_DIR}/xasl_regu_alloc.hpp
-  ${PARSER_DIR}/predication/parser_expr_compiler.hpp
   )
 
 set(OPTIMIZER_SOURCES
@@ -271,6 +269,9 @@ set(QUERY_SOURCES
   ${QUERY_DIR}/query_cl.c
   ${QUERY_DIR}/query_dump.c
   ${QUERY_DIR}/expression/compare.cc
+  ${QUERY_DIR}/expression/cast.cc
+  ${QUERY_DIR}/expression/compiled_pred_expr_interpreter.cpp
+  ${QUERY_DIR}/expression/parser_expr_compiler.cpp
   ${QUERY_DIR}/query_evaluator.c
   ${QUERY_DIR}/query_executor.c
   ${QUERY_DIR}/query_manager.c
@@ -293,7 +294,10 @@ set(QUERY_SOURCES
   ${QUERY_DIR}/xasl_to_stream.c
   )
 set(QUERY_HEADERS
-  ${QUERY_DIR}/expression/compare.h
+  ${QUERY_DIR}/expression/compare.hpp
+  ${QUERY_DIR}/expression/cast.hpp
+  ${QUERY_DIR}/expression/compiled_pred_expr_interpreter.hpp
+  ${QUERY_DIR}/expression/parser_expr_compiler.hpp
   ${QUERY_DIR}/query_aggregate.hpp
   ${QUERY_DIR}/query_hash_scan.h
   ${QUERY_DIR}/query_analytic.hpp
@@ -603,7 +607,6 @@ SET_SOURCE_FILES_PROPERTIES(
   ${PARSER_DIR}/semantic_check.c
   ${PARSER_DIR}/type_checking.c
   ${PARSER_DIR}/xasl_generation.c
-  ${PARSER_DIR}/predication/parser_expr_compiler.cpp
   ${PARSER_DIR}/scanner_support.c
   ${PARSER_DIR}/view_transform.c
   ${OPTIMIZER_SOURCES}
